리팩토링 : 상품 데이터베이스 변경으로 인한 개별 상품 가져오기 수정
