+++ src/main/java/com/project/onlinemarketplaceservice/dto/ProductDetailDto.java
@@ -1,5 +1,7 @@
 package com.project.onlinemarketplaceservice.dto;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,12 +11,14 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class InsertProductDto {
+public class ProductDetailDto {
 
+    private Long productId;
     private BaseProductDto baseProductDto;
     private String description;
     private String detailedImage1;
     private String detailedImage2;
     private String detailedImage3;
+    private LocalDateTime publishedAt;
 
 }
