+++ src/test/java/com/project/onlinemarketplaceservice/MarcketplaceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.project.onlinemarketplaceservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MarcketplaceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
