--- src/test/java/com/project/smartstore/SmartStoreApplicationTests.java
@@ -1,13 +1,0 @@
-package com.project.smartStore;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SmartStoreApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
